x86: grow rootfs_data partition before format
The non-default f2fs-tools can expand a f2fs only when not in use.
/overlay is only unmounted while in failsafe, which also prevents the
use of any installed packages. Today, the easier way is to resize the
partition after installation is resizing the partition and use failsafe
to reformat the /overlay, requiring physical device access. And, after
an upgrade, the partition is reverted.

This patch adds a grow_rootfspart shellscript partitioner that expands
the seconds MBR or GPT partition until the end of the device. It can
run directly to an img file or from the installed device. It only requires
standard Linux commands already available on default x86 images. After
the partition resizes, it will refresh partition information in kernel.

For GPT, it will always touch the disk at least once as the current
generated images do not use all the available space and it also misses
the GPT backup header at the end of the disk.

preinit/43_grow_rootfspart will run it when it does not detect an f2fs/ext4
filesystem at the overlay position and only when not in failsafe. That
condition is also met during the first boot after an upgrade. mount_root
will format rootfs_data alter that.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>