generic: 5.15: backport revert patch for DSA mtu initialization
The upstream patch delegated to slave port the job of settings master
port MTU. However, it depends on DSA driver implementing the functions
to redefine the MTU for a specific port. When that optional feature was
missing, the cpu port MTU was never updated. It is not enough to drop
765-3-net-next-net-dsa-stop-updating-master-MTU-from-master.c.patch
as the changes in the lock sequence might result in a deadlock.

Link: https://lore.kernel.org/netdev/20220331132854.1395040-1-vladimir.oltean@nxp.com/
Fixes: 327b6dbd98 (generic: 5.15: backport mdio improvement patch for qca8k)
Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>