ath79: convert TP-Link TL-WRD3600,TL-WRD4300 to DSA
Both normal and failsafe are working as expected.
tl-wdr4300-v1-il was migrated but not tested in real HW.

LEDs are using link_10, link_100 and link_1000 instead of plain link as
rx,tx does not work with link in current kernel. It looks like it is
enabling offload without actually supporting link.

Added qca8k-0.0 to led migration exclusion as the prefix comes from the
driver.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>