ath79: convert TP-Link TL-WRD3600,TL-WRD4300 to DSA
Both normal and failsafe are working as expected.
tl-wdr4300-v1-il was migrated but not tested in real HW.

WAN/LAN leds are now visible in sysfs but left unconfigured as it was
not possible to reproduce the default behavior configuring sysfs.

The link indication is not working together with rx/tx but it works if used
alone. The netdev also fails to monitor other device than the default
one. Setting interval is failing with "Invalid argument", visible when
led service is started (including during boot).

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>