ath79: migrate tl-wr2543nd to DSA (WIP)
This device uses a RTL8367R (RTL8367 family) device. The driver
rtl8365mb was adapted to support it but it is still a WIP.

The RTL8367 family lacks transparent VLAN feature, which is needed for
disabling VLAN filtering for a bridge or a standalone port. The only
found solution was to add CPU ports to all possible VLANs. This is
currently taking 10s.

The code still needs some cleanups and missing features:
1) LEDs still not supported
2) supported tags / MTU not tested
3) MIB might need to be updated

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>