ath79: convert TP-Link TL-WR1043ND v1 to DSA (BROKEN)
Drop swconfig packages and include the DSA ones. This is the first one
to use the upstraem Realtek DSA switch driver.

It is working as far as possible (32M RAM), both in normal and failsafe
mode. Switch LEDs required upstream fix.

FIXME: test in real HW failed. DO NOT FLASH IT YET

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>