generic: hwmon: (WIP) add initialization commmands to lm63
Some lm63 registers are not touched by the driver as it is normally
expected that the BIOS or bootloader initialize it before the OS is
loaded.  However, for devices like D-Link dgs-1210-28p-f and
dgs-1210-28mp-f, the lm63 initialization is done by the OEM firmware.

This commit adds an option to inject values into any lm63 register,
allowing openwrt to mimic the OEM sequence.

TODO: the userland solution is still enabled. You have to manually
disable fan_ctrl.sh to test it.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>